refactor: PathPool -> StaticResourcePathPool, ApiPathPool 로 분리
