+++ src/main/java/util/ApiPathPool.java
@@ -13,17 +13,16 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.concurrent.ConcurrentHashMap;
 
-public class PathPool {
-    private static final Logger logger = LoggerFactory.getLogger(PathPool.class);
+public class ApiPathPool {
+    private static final Logger logger = LoggerFactory.getLogger(ApiPathPool.class);
     private final ConcurrentHashMap<String, ConcurrentHashMap<HttpMethod, Method>> methodMap = new ConcurrentHashMap<>();
     private final ConcurrentHashMap<String, Object> classMap = new ConcurrentHashMap<>();
 
-    private final ConcurrentHashMap<String, Handler> staticResourceMap = new ConcurrentHashMap<>();
-    private static final PathPool instance;
+    private static final ApiPathPool instance;
 
     static {
         try {
-            instance = new PathPool();
+            instance = new ApiPathPool();
         } catch (InvocationTargetException e) {
             throw new RuntimeException(e);
         } catch (InstantiationException e) {
@@ -35,9 +34,8 @@ public class PathPool {
         }
     }
 
-    private PathPool() throws InvocationTargetException, InstantiationException, IllegalAccessException, NoSuchMethodException {
+    private ApiPathPool() throws InvocationTargetException, InstantiationException, IllegalAccessException, NoSuchMethodException {
         initApiPath();
-        initStaticResourcePath();
     }
 
     private void initApiPath() throws InvocationTargetException, InstantiationException, IllegalAccessException, NoSuchMethodException {
@@ -64,12 +62,6 @@ public class PathPool {
         classMap.put("/user/logout", userHandler);
     }
 
-    private void initStaticResourcePath() {
-        staticResourceMap.put("/", new HomeHandler());
-        staticResourceMap.put("/main", new MainHandler());
-        staticResourceMap.put("/mypage", new MypageHandler());
-    }
-
     public boolean isAvailable(HttpMethod method, String path) {
         if (!classMap.containsKey(path)) {
             return false;
@@ -80,7 +72,7 @@ public class PathPool {
         return true;
     }
 
-    public static PathPool getInstance() {
+    public static ApiPathPool getInstance() {
         return instance;
     }
 
@@ -91,9 +83,4 @@ public class PathPool {
     public Object getClass(String path) {
         return classMap.get(path);
     }
-
-    public Handler getHandler(String path) {
-        return staticResourceMap.get(path);
-    }
-
 }
